requirements.txt

streamlit
pandas
numpy
requests
scikit-learn
joblib
plotly
sqlalchemy
psycopg2-binary
spacy
python-dotenv
pytest
